+++ test/byFunctions/IexecAPI/workOrderCallback.js
@@ -9,7 +9,6 @@ var WorkOrder = artifacts.require("./WorkOrder.sol");
 var IexecLib = artifacts.require("./IexecLib.sol");
 var Marketplace = artifacts.require("./Marketplace.sol");
 var IexecAPI = artifacts.require("./IexecAPI.sol");
-var CallbackProof = artifacts.require("./CallbackProof.sol");
 
 const Promise = require("bluebird");
 const fs = require("fs-extra");
@@ -47,7 +46,6 @@ contract('IexecHub', function(accounts) {
   let aAppHubInstance;
   let aDatasetHubInstance;
   let aMarketplaceInstance;
-  let aCallbackProofInstance;
 
   //specific for test :
   let workerPoolAddress;
@@ -150,12 +148,6 @@ contract('IexecHub', function(accounts) {
     assert.strictEqual(balances[6].toNumber(), 1000, "1000 nRLC here");
 
     // INIT SMART CONTRACTS BY marketplaceCreator
-    aCallbackProofInstance = await CallbackProof.new({
-      from: marketplaceCreator
-    });
-    console.log("aCallbackProofInstance.address is ");
-    console.log(aCallbackProofInstance.address);
-
     aWorkerPoolHubInstance = await WorkerPoolHub.new({
       from: marketplaceCreator
     });
@@ -312,7 +304,7 @@ contract('IexecHub', function(accounts) {
   it("workOrderCallback_01: test workOrderCallback from a smart contract", async function() {
 
 
-    aIexecAPIInstance = await IexecAPI.new(aIexecHubInstance.address, aCallbackProofInstance.address,aRLCInstance.address, {
+    aIexecAPIInstance = await IexecAPI.new(aIexecHubInstance.address, aMarketplaceInstance.address,aRLCInstance.address, {
       from: iExecCloudUser
     });
     console.log("aIexecAPIInstance created " + aIexecAPIInstance.address);
@@ -489,7 +481,7 @@ contract('IexecHub', function(accounts) {
     assert.strictEqual(checkBalance[1].toNumber(), 0, "check stake locked of the scheduleProvider");
 
 
-    txMined = await aCallbackProofInstance.workOrderCallback(woid, "aStdout", "aStderr", "anUri", {
+    txMined = await aMarketplaceInstance.workOrderCallback(woid, "aStdout", "aStderr", "anUri", {
       from: scheduleProvider,
       gas: constants.AMOUNT_GAS_PROVIDED
     });